:white_check_mark: Test: Add Evm test
