+++ packages/evm/package.json
@@ -65,7 +65,8 @@
     "@ethereumjs/evm": "^3.0.0",
     "@tevm/blockchain": "workspace:^",
     "@tevm/common": "workspace:^",
-    "@tevm/predeploys": "workspace:^"
+    "@tevm/predeploys": "workspace:^",
+    "@tevm/utils": "workspace:1.0.0-next.40"
   },
   "devDependencies": {
     "@tevm/state": "workspace:^",
