+++ packages/evm/src/index.ts
@@ -1,16 +1,17 @@
 export { createEvm } from './createEvm.js'
-export { Evm, type EVMOpts } from './Evm.js'
+export { Evm } from './Evm.js'
+export { type EVMOpts } from './EvmOpts.js'
 export type { CreateEvmOptions } from './CreateEvmOptions.js'
 export type { CustomPrecompile } from './CustomPrecompile.js'
 export {
-	EVMErrorMessage as EvmErrorMessage,
-	EvmError,
-	EOF as Eof,
-	Message as EthjsMessage,
-	getActivePrecompiles,
-	type PrecompileInput,
-	type InterpreterStep,
-	type ExecResult,
-	type EVMResult as EvmResult,
-	type EVMRunCallOpts as EvmRunCallOpts,
+  EVMErrorMessage as EvmErrorMessage,
+  EvmError,
+  EOF as Eof,
+  Message as EthjsMessage,
+  getActivePrecompiles,
+  type PrecompileInput,
+  type InterpreterStep,
+  type ExecResult,
+  type EVMResult as EvmResult,
+  type EVMRunCallOpts as EvmRunCallOpts,
 } from '@ethereumjs/evm'
